:lipstick: add and style the slider button
