+++ apps/elewa-website/src/app/app.component.html
@@ -1,5 +1,7 @@
 
 <!-- Render the elewa website header component before the router outlet and at the top of the main app component so that the header can be displayed at the top of every page / route of the main application -->
-<elewa-website-header></elewa-website-header>
+<!-- <elewa-website-header></elewa-website-header> -->
 
 <router-outlet></router-outlet>
+
+<elewa-website-elewa-slider-button [data]="sliderButtonData" (click)="handleClick()"></elewa-website-elewa-slider-button>
