--- apps/elewa-website/src/app/elewa-slider-button/elewa-slider-button.component.html
@@ -1,1 +1,0 @@
-<p>elewa-slider-button works!</p>
