--- apps/elewa-website/src/app/elewa-slider-button/elewa-slider-button.component.ts
@@ -1,10 +1,0 @@
-import { Component } from '@angular/core';
-
-@Component({
-  selector: 'elewa-website-elewa-slider-button',
-  templateUrl: './elewa-slider-button.component.html',
-  styleUrls: ['./elewa-slider-button.component.scss'],
-})
-export class ElewaSliderButtonComponent {
-
-}
