+++ libs/elements/layout/buttons/src/lib/buttons.module.ts
@@ -1,7 +1,8 @@
 import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
+import { ElewaSliderButtonComponent } from './elewa-slider-button/elewa-slider-button.component';
 
 @NgModule({
-  imports: [CommonModule],
+  declarations: [ElewaSliderButtonComponent],
+  exports: [ElewaSliderButtonComponent]
 })
 export class ButtonsModule {}
