+++ libs/elements/layout/buttons/src/lib/elewa-slider-button/elewa-slider-button.component.html
@@ -1,0 +1,4 @@
+<button class="elewa-slider-button">
+    <span class="button-text">Unlock your potential</span>
+    <i class="material-icons">arrow_right_alt</i>
+</button>
