+++ package.json
@@ -28,6 +28,7 @@
     "@nguniversal/express-engine": "^16.1.1",
     "@nrwl/angular": "16.6.0",
     "express": "^4.15.2",
+    "material-icons": "^1.13.9",
     "rxjs": "7.8.1",
     "tslib": "^2.3.0",
     "zone.js": "0.13.1"
